requirements.txt
joblib
numpy >= 1.10.4
opencv-contrib-python
Pillow
scikit-image
scikit-learn >= 0.19.0
scipy
torch >= 1.0
torchvision >= 0.2.1